feat(feature:savings-account): Migrate savings-account module to CMP
