+++ feature/savings-account/src/commonMain/kotlin/org/mifos/mobile/feature/savingsaccount/component/AccountTypeItemIndicator.kt
@@ -7,7 +7,7 @@
  *
  * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
  */
-package org.mifos.mobile.feature.savingsaccount.screen.component
+package org.mifos.mobile.feature.savingsaccount.component
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
