+++ gradle/libs.versions.toml
@@ -324,7 +324,6 @@ moko-permission-compose = { group = "dev.icerock.moko", name = "permissions-comp
 window-size = { group = "dev.chrisbanes.material3", name = "material3-window-size-class-multiplatform", version.ref = "windowsSizeClass" }
 androidx-room-common = { group = "androidx.room", name = "room-common", version.ref = "roomCommonVersion" }
 material = { group = "com.google.android.material", name = "material", version.ref = "materialVersion" }
-androidx-ui-android = { group = "androidx.compose.ui", name = "ui-android", version.ref = "uiAndroidVersion" }
 
 [bundles]
 androidx-compose-ui-test = [
